fix: remove target_correlation from parameters
use the correlation matrix of the passed space instead
