+++ experiment_design/scorers.py
@@ -4,7 +4,6 @@ import numpy as np
 from scipy.spatial.distance import pdist
 
 from experiment_design.variable import ParameterSpace
-from experiment_design.variable.space import VariableCollection
 
 
 class Scorer(Protocol):
@@ -22,7 +21,7 @@ class Scorer(Protocol):
 class ScorerFactory(Protocol):
     def __call__(
         self,
-        variables: VariableCollection,
+        space: ParameterSpace,
         sample_size: int,
         old_sample: np.ndarray | None = None,
     ) -> Scorer:
@@ -34,8 +33,6 @@ class MaxCorrelationScorerFactory:
     """
     A scorer factory for the maximum absolute correlation error between sampling points.
 
-    :param target_correlation: A symmetric matrix with shape (len(variables), len(variables)), representing the linear
-        dependency between the dimensions.
     :param local: If True, any points in the old_sample will be ignored, that fall outside the finite bounds of the
         provided variables. Has no effect if old_sample is None.
     :param eps: a small positive value to improve the stability of the log operation.
@@ -44,14 +41,9 @@ class MaxCorrelationScorerFactory:
 
     def __init__(
         self,
-        target_correlation: float | np.ndarray = 0,
         local: bool = True,
         eps: float = 1e-2,
     ) -> None:
-        if np.max(np.abs(target_correlation)) > 1:
-            raise ValueError("Correlations should be in the interval [-1,1].")
-        self.target_correlation = target_correlation
-
         self.local = local
 
         if eps < 0:
@@ -60,7 +52,7 @@ class MaxCorrelationScorerFactory:
 
     def __call__(
         self,
-        variables: VariableCollection,
+        space: ParameterSpace,
         sample_size: int,
         old_sample: np.ndarray | None = None,
     ) -> Scorer:
@@ -68,7 +60,7 @@ class MaxCorrelationScorerFactory:
         Creates a scorer, that computes the maximum absolute correlation error between the candidate samples
         and the target_correlation.
 
-        :param variables: Dimensions of the design space.
+        :param space: Dimensions of the design space.
         :param sample_size: number of candidate points to be scored.
         :param old_sample: If passed, represents the matrix of points in an older design of experiments with shape
             (old_sample_size, len(variables)). Depending on self.local, some or all of these will be appended to the
@@ -77,14 +69,13 @@ class MaxCorrelationScorerFactory:
             smaller than self.eps, it returns negative exp(maximum absolute correlation error - 1) instead.
         """
 
-        target_correlation = create_correlation_matrix(
-            self.target_correlation, num_variables=len(variables)
-        )
-        handler = create_old_doe_handler(variables, old_sample, local=self.local)
+        handler = create_old_doe_handler(space, old_sample, local=self.local)
 
         def _scorer(doe: np.ndarray) -> float:
             error = np.max(
-                np.abs(np.corrcoef(handler(doe), rowvar=False) - target_correlation)
+                np.abs(
+                    np.corrcoef(handler(doe), rowvar=False) - space.correlation_matrix
+                )
             )
             if error > self.eps:
                 return -np.exp(error + 1)
@@ -113,14 +104,14 @@ class PairwiseDistanceScorerFactory:
 
     def __call__(
         self,
-        variables: VariableCollection,
+        space: ParameterSpace,
         sample_size: int,
         old_sample: np.ndarray | None = None,
     ) -> Scorer:
         """
         Create a scorer, that computes the minimum pairwise distance between sampling points.s
 
-        :param variables: Dimensions of the design space.
+        :param space: Dimensions of the design space.
         :param sample_size: number of candidate points to be scored.
         :param old_sample: If passed, represents the matrix of points in an older design of experiments with shape
             (old_sample_size, len(variables)). Depending on self.local, some or all of these will be appended to the
@@ -128,9 +119,9 @@ class PairwiseDistanceScorerFactory:
         :return: a scorer that returns the log minimum pairwise distance divided by the log max distance,
         """
 
-        handler = create_old_doe_handler(variables, old_sample, local=self.local)
+        handler = create_old_doe_handler(space, old_sample, local=self.local)
         bin_diagonal_length = calculate_equidistant_bin_diagonal_length(
-            variables, sample_size
+            space, sample_size
         )
 
         def _scorer(doe: np.ndarray) -> float:
@@ -162,13 +153,13 @@ class WeightedSumScorerFactory:
 
     def __call__(
         self,
-        variables: VariableCollection,
+        space: ParameterSpace,
         sample_size: int,
         old_sample: np.ndarray | None = None,
     ) -> Scorer:
 
         scorers = [
-            factory(variables, sample_size, old_sample=old_sample)
+            factory(space, sample_size, old_sample=old_sample)
             for factory in self.scorer_factories
         ]
 
@@ -181,7 +172,6 @@ class WeightedSumScorerFactory:
 
 
 def create_default_scorer_factory(
-    target_correlation: float | np.ndarray = 0.0,
     distance_score_weight: float = 0.9,
     correlation_score_weight: float = 0.1,
     local_correlation: bool = True,
@@ -191,18 +181,13 @@ def create_default_scorer_factory(
     Create a scorer factory, which is a weighted sum of maximum correlation error and
     minimum pairwise distance scorers
 
-    :param target_correlation: A symmetric matrix with shape (len(variables), len(variables)), representing the linear
-        dependency between the dimensions. If a float, all non-diagonal entries of the unit matrix will be set to this
-        value.
     :param distance_score_weight: Weight of the minimum pairwise distance score.
     :param correlation_score_weight: Weight of the maximum correlation error score.
     :param local_correlation: Controls the local attribute of the MaxCorrelationScorerFactory.
     :param local_pairwise_distance: Controls the local attribute of the PairwiseDistanceScorerFactory
     :return: WeightedSumScorerFactory instance.
     """
-    corr_scorer_factory = MaxCorrelationScorerFactory(
-        target_correlation=target_correlation, local=local_correlation
-    )
+    corr_scorer_factory = MaxCorrelationScorerFactory(local=local_correlation)
     dist_scorer_factory = PairwiseDistanceScorerFactory(local=local_pairwise_distance)
     return WeightedSumScorerFactory(
         scorer_factories=[corr_scorer_factory, dist_scorer_factory],
@@ -211,34 +196,15 @@ def create_default_scorer_factory(
 
 
 def calculate_equidistant_bin_diagonal_length(
-    variables: VariableCollection, sample_size: int
+    space: ParameterSpace, sample_size: int
 ) -> float:
     """Calculate the length of the diagonal of equidistant bins the (Euclidean) design space"""
-    if not isinstance(variables, ParameterSpace):
-        variables = ParameterSpace(variables)
-    lower, upper = variables.lower_bound, variables.upper_bound
+    lower, upper = space.lower_bound, space.upper_bound
     return np.linalg.norm((np.array(upper) - np.array(lower)) / sample_size)
 
 
-def create_correlation_matrix(
-    target_correlation: float | np.ndarray = 0.0,
-    num_variables: int | None = None,
-) -> np.ndarray:
-    """Create a correlation matrix from the target correlation in case it is a float"""
-    if not np.isscalar(target_correlation):
-        return target_correlation
-    if not num_variables:
-        raise ValueError(
-            "num_variables have to be passed if the target_correlation is a scalar."
-        )
-    return (
-        np.eye(num_variables) * (1 - target_correlation)
-        + np.ones((num_variables, num_variables)) * target_correlation
-    )
-
-
 def create_old_doe_handler(
-    variables: VariableCollection,
+    space: ParameterSpace,
     old_sample: np.ndarray | None = None,
     local: bool = False,
 ) -> Callable[[np.ndarray], np.ndarray]:
@@ -247,7 +213,7 @@ def create_old_doe_handler(
     append some or all of the points from an old design of experiments (DoE) for
     including them in the scoring.
 
-    :param variables: Dimensions of the design space. Only relevant if local is set to True.
+    :param space: Dimensions of the design space. Only relevant if local is set to True.
     :param old_sample: Matrix of points with shape=(sample_size, len(variables)) in the old DoE. If None, this returns
         a no-op function.
     :param local: If True, only include the points from the old_doe, that fall between the finite local bounds of the
@@ -263,14 +229,12 @@ def create_old_doe_handler(
         # Append every point in the old doe
         return lambda x: np.append(old_sample, x, axis=0)
 
-    old_sample = select_local(old_sample, variables)
+    old_sample = select_local(old_sample, space)
     return lambda x: np.append(old_sample, x, axis=0)
 
 
-def select_local(samples: np.ndarray, variables: VariableCollection) -> np.ndarray:
+def select_local(samples: np.ndarray, space: ParameterSpace) -> np.ndarray:
     """Select and return samples that fall within the finite bounds of the variables"""
-    if not isinstance(variables, ParameterSpace):
-        variables = ParameterSpace(variables)
-    lower, upper = variables.lower_bound[None, :], variables.upper_bound[None, :]
+    lower, upper = space.lower_bound[None, :], space.upper_bound[None, :]
     local_mask = np.logical_and((samples >= lower).all(1), (samples <= upper).all(1))
     return samples[local_mask]
